interview-service-ai/AudioAnswerConverterService/requirements.txt
kafka-python==2.2.15
python-dotenv==1.1.1
mongoengine==0.29.1
requests==2.32.5
faster-whisper==1.2.1